Rudimentarily split webcl-validator to a lib and a driver binary
The current API is basically a C++ class that implements main().

The plan is to move the command line specific stuff to the driver binary
and leave the validator guts in the library, behind the public library
API. But that's gonna require some heavy refactoring.
