+++ driver/main.cpp
@@ -1,0 +1,45 @@
+/*
+** Copyright (c) 2013 The Khronos Group Inc.
+**
+** Permission is hereby granted, free of charge, to any person obtaining a
+** copy of this software and/or associated documentation files (the
+** "Materials"), to deal in the Materials without restriction, including
+** without limitation the rights to use, copy, modify, merge, publish,
+** distribute, sublicense, and/or sell copies of the Materials, and to
+** permit persons to whom the Materials are furnished to do so, subject to
+** the following conditions:
+**
+** The above copyright notice and this permission notice shall be included
+** in all copies or substantial portions of the Materials.
+**
+** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
+** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
+*/
+
+#include <wclv/wclv.h>
+
+#include <iostream>
+#include <set>
+#include <string>
+
+int main(int argc, char const* argv[])
+{
+    std::set<std::string> help;
+    help.insert("-h");
+    help.insert("-help");
+    help.insert("--help");
+
+    if ((argc == 1) || ((argc == 2) && help.count(argv[1]))) {
+        std::cerr << "Usage: " << argv[0] << " input.cl [clang-options]" << std::endl;
+        return EXIT_FAILURE;
+    }
+
+    WebCLValidator validator(argc, argv);
+
+    return validator.run();
+}
