+++ lib/WebCLCommon.hpp
