+++ lib/WebCLConfiguration.cpp
