+++ lib/WebCLConfiguration.hpp
