+++ lib/WebCLHelper.cpp
