+++ lib/WebCLReporter.cpp
