+++ lib/WebCLRewriter.hpp
