+++ lib/general.cl
