+++ lib/kernel.cl
