+++ lib/stripcr.py
